backend/pyproject.toml
[project]
name = "wordpress-agent"
version = "0.3.0"
description = "AI-powered WordPress site builder agent with Google Drive integration"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "beautifulsoup4>=4.14.0",
    "fastapi[standard]>=0.133.0",
    "httpx>=0.28.0",
    "python-dotenv>=1.2.0",
    # Settings management
    "pydantic-settings>=2.7.0",
    # Database
    "sqlalchemy[asyncio]>=2.0.36",
    "asyncpg>=0.30.0",
    "alembic>=1.14.0",
    # Google OAuth & Drive API
    "google-auth>=2.37.0",
    "google-auth-oauthlib>=1.2.1",
    "google-auth-httplib2>=0.2.0",
    "google-api-python-client>=2.159.0",
    # LangChain + LangGraph (agentic core)
    "langchain-core>=1.2.8,<2",
    "langchain-groq>=1.1.0,<2",
    "langchain-openai>=1.0.0,<2",  # For GLM-5 (OpenAI-compatible)
    "langchain-google-genai>=2.0.0,<3",  # For Gemini
    "langgraph>=1.0.0,<2",
    "langgraph-checkpoint-sqlite>=2.0.0,<3",  # For state persistence
    # Authentication
    "pyjwt>=2.11.0",
    "bcrypt>=5.0.0",
    "passlib>=1.7.4",
    "email-validator>=2.3.0",
    # Code quality and testing
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "hypothesis>=6.122.0",
    "pytest-mock>=3.14.0",
    "flake8>=7.1.0",
    "black>=24.10.0",
    "mypy>=1.13.0",
    "bandit>=1.8.0",
    "tomli>=2.2.0",
]

[tool.black]
line-length = 100
target-version = ['py314']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
  | alembic/versions
)/
'''

[tool.mypy]
python_version = "3.14"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
follow_imports = "normal"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "alembic.*"
ignore_errors = true

[tool.pytest.ini_options]
minversion = "8.0"
addopts = "-ra -q --strict-markers --cov=app --cov-report=term-missing --cov-report=html"
testpaths = ["tests"]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"

[tool.coverage.run]
source = ["app"]
omit = [
    "*/tests/*",
    "*/alembic/*",
    "*/__pycache__/*",
    "*/venv/*",
    "*/.venv/*",
]

[tool.coverage.report]
precision = 2
show_missing = true
skip_covered = false

[tool.bandit]
exclude_dirs = ["/tests", "/alembic"]
skips = ["B101", "B601"]
